chore: restructure views
