+++ src/ManagementPortal/Api/Views/Applications/EditApplication.cshtml
@@ -1,4 +1,3 @@
-@using ManagementPortal.Api.Views.Applications.Shared
 @model EditApplicationModel
 
 @{
@@ -42,7 +41,7 @@
 
     <div class="px-6 py-5">
         <form method="post">
-            <partial name="Shared/_ApplicationForm" model="formModel" />
+            <partial name="_ApplicationForm" model="formModel" />
         </form>
     </div>
 </div>
