+++ src/ManagementPortal/Api/Views/Applications/EditApplication.cshtml.cs
@@ -1,5 +1,3 @@
-using ManagementPortal.Api.Views.Applications.Shared;
-
 namespace ManagementPortal.Api.Views.Applications;
 
 public class EditApplicationModel: ApplicationFormModel;
