+++ src/ManagementPortal/Api/Views/Applications/Index.cshtml
@@ -8,7 +8,7 @@
 <div class="bg-white shadow overflow-hidden sm:rounded-lg">
     <div class="flex justify-between items-center px-6 py-4 border-b border-gray-200">
         <h1 class="text-2xl font-semibold text-gray-900">OIDC Applications</h1>
-        <a asp-action=""
+        <a asp-action="CreateNewApplication"
            class="inline-flex items-center px-4 py-2 border border-transparent rounded-md shadow-sm text-sm font-medium text-white bg-indigo-600 hover:bg-indigo-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500">
             <svg class="-ml-1 mr-2 h-5 w-5" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20" fill="currentColor">
                 <path fill-rule="evenodd"
