+++ src/ManagementPortal/Api/Views/Applications/_DynamicListInput.cshtml
@@ -1,4 +1,4 @@
-@model _DynamicListInput
+@model DynamicListInput
 @inject Microsoft.AspNetCore.Antiforgery.IAntiforgery Antiforgery
 
 <div>
